# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Boardman API
  description: P2P bet escrow platform API
  version: '1.0'
servers:
  - url: http://localhost:3002
    description: Local development
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: boardman_token
  schemas:
...
    FundWalletDto:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          minimum: 100
          description: Amount in Naira (min 100)
          example: 500
    Transaction:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [DEPOSIT, WITHDRAWAL, ESCROW_LOCK, ESCROW_RELEASE, REFUND, FEE]
        amount:
          type: number
        betId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    WalletResponse:
      type: object
      properties:
        id:
          type: string
        walletBalance:
          type: number
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
    FundInitiateResponse:
      type: object
      properties:
        paymentUrl:
          type: string
          example: 'https://pay.interswitch.com/pay'
        reference:
          type: string
          example: REF123

    # ── Bets ──────────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string
paths:
...
  /wallet:
    get:
      tags: [Wallet]
      summary: Get wallet balance and last 50 transactions
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Wallet data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /wallet/fund/initiate:
    post:
      tags: [Wallet]
      summary: Initiate a wallet top-up via Interswitch Pay Bill
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FundWalletDto'
      responses:
        '201':
          description: Payment initiated — redirect user to paymentUrl
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FundInitiateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Interswitch gateway error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...